Adding the missing original x-men from 1963–1966
I get it Sam, Wolverine is your favorite. 
But i find you lack of respect for the original members in Charles Xavier´s X-Men from the silver age of Marvel comics disturbing.
I kept your pressios Wolverine in the list but this list is not complete without the original members!